move the logic for storring external account ops to activity.ts
